[iac] Separate GitHub credential audit concerns
