+++ infra/pulumi/github/__main__.py
@@ -13,19 +13,24 @@ import pulumi
 from iac.github.credentials import credential_manifest
 from iac.github.resources import credential_resource_plans, register_credentials
 
-config = pulumi.Config()
-manifest = credential_manifest(
-    schema_version=config.require_int("schemaVersion"),
-    organization=config.require("organization"),
-    repositories=cast(list[str], config.require_object("repositories")),
-    credentials=cast(list[dict], config.require_object("credentials")),
-)
-provider_owner = pulumi.Config("github").require("owner")
-if provider_owner != manifest.organization:
-    raise ValueError(
-        f"github:owner {provider_owner!r} does not match marin-github:organization {manifest.organization!r}"
-    )
 
-plans = credential_resource_plans(manifest)
-register_credentials(manifest)
-pulumi.export("credential_count", len(plans))
+def main() -> None:
+    config = pulumi.Config()
+    manifest = credential_manifest(
+        schema_version=config.require_int("schemaVersion"),
+        organization=config.require("organization"),
+        repositories=cast(list[str], config.require_object("repositories")),
+        credentials=cast(list[dict], config.require_object("credentials")),
+    )
+    provider_owner = pulumi.Config("github").require("owner")
+    if provider_owner != manifest.organization:
+        raise ValueError(
+            f"github:owner {provider_owner!r} does not match marin-github:organization {manifest.organization!r}"
+        )
+
+    plans = credential_resource_plans(manifest)
+    register_credentials(manifest)
+    pulumi.export("credential_count", len(plans))
+
+
+main()
